# api/swagger/swagger.yaml
basePath: /api
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  dto.ConfirmPasswordResetRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.OrgSummaryResponse:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.RequestPasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  dto.UserResponse:
    properties:
      active_org_id:
        type: string
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      org:
        $ref: '#/definitions/dto.OrgSummaryResponse'
      role:
        type: string
      workspaces:
        items:
          $ref: '#/definitions/dto.WorkspaceResponse'
        type: array
    type: object
...
  dto.WorkspaceResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_owner:
        type: boolean
      name:
        type: string
      role:
        type: string
      slug:
        type: string
    type: object
  transport.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  transport.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Self-hosted project management tool API
  title: Plume API
  version: "1.0"
paths:
  /account:
    patch:
      consumes:
      - application/json
      description: Changes the caller's display name (synced across all workspaces)
      parameters:
      - description: Profile update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Update own profile
      tags:
      - account
  /account/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an avatar image for the caller's account (synced across
        all workspaces)
      parameters:
      - description: Image file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Upload avatar
      tags:
      - account
  /account/change-password:
    post:
      consumes:
      - application/json
      description: Changes the caller's password (current password required); revokes
        sessions on success
      parameters:
      - description: Password change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transport.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Change password
      tags:
      - account
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, sets HttpOnly cookie
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /auth/me:
    get:
      description: Returns the authenticated user's profile with org info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Get current user
      tags:
      - auth
  /auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      description: Completes a password reset with the token from the reset link and
        a new password
      parameters:
      - description: Token + new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConfirmPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transport.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Confirm password reset
      tags:
      - auth
  /auth/password-reset/request:
    post:
      consumes:
      - application/json
      description: Generates a password reset token (logged server-side; no SMTP configured
        for self-hosters)
      parameters:
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RequestPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transport.SuccessResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/password-reset/validate:
    get:
      description: Checks if a password-reset token is valid (exists, not expired,
        not used) without consuming it. Returns 200 with valid=true or 400 with valid=false.
      parameters:
      - description: Reset token from the email
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"valid\": true}'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: '{\"valid\": false}'
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Validate password reset token
      tags:
      - auth
...
  /auth/sessions/{id}:
    delete:
      description: Revokes one of the caller's other sessions ("log out" that device).
        The current session cannot be revoked here; use /auth/logout.
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transport.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/transport.ErrorResponse'
      summary: Revoke a session
      tags:
      - auth
...
swagger: "2.0"